[Feat] NPC 상호작용 추가
NPC에 다가가서 Z키를 누르면 상호작용 디버그 로그 출력
