+++ Assets/MSK/MSKScripts/IInteractable.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public interface IInteractable
+{
+    void Interact();
+}
