+++ Assets/MSK/MSKScripts/MSKNPCCollision.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5ea7669f522983240a6f6e3969db8010
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
