+++ Assets/MSK/MSKScripts/MSKPCcontroller.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b3f684e350513a94fbec8072fb522758
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
