+++ Assets/MSK/MSKScripts/NPCController.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class NPCController : MonoBehaviour, IInteractable
+{
+    public void Interact() 
+    {
+        Debug.Log("NPC Interact");
+    }
+
+}
