+++ Assets/MSK/MSKScripts/NPCController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 02ef73625943f2d498f113e21dd28dd7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
